payment interface moved into it's own folder
